Move exported test classes under `org.apache.logging.log4j.core.test`
