+++ log4j-slf4j2-impl/src/test/java/org/apache/logging/slf4j/Log4j2_1482_Slf4jTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.logging.slf4j;
 
-import org.apache.logging.log4j.core.layout.Log4j2_1482_Test;
+import org.apache.logging.log4j.core.test.layout.Log4j2_1482_Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
